requirements.txt
# YouTube trending tool
google-api-python-client>=2.0.0
python-dotenv>=1.0.0

# SaaS web app
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
jinja2>=3.1.0
python-multipart>=0.0.9
starlette>=0.36.0
itsdangerous>=2.1.0

# Database
sqlalchemy>=2.0.0

# Auth
bcrypt>=4.0.0

# Payments
stripe>=8.0.0

# Scheduler
apscheduler>=3.10.0
requests>=2.0.0